patch change

+++ src/ripple/app/tx/impl/Change.cpp
@@ -528,7 +528,7 @@ Change::activateXahauGenesis()
     }
 
     {
-        ripple::STArray hooks{sfHooks, genesis_hooks.size()};
+        ripple::STArray hooks{sfHooks, static_cast<int>(genesis_hooks.size())};
         int hookCount = 0;
 
         for (auto const& [hookOn, wasmBytes, params] : genesis_hooks)
